# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RBAC 管理后台接口，覆盖登录认证、菜单/API 权限、角色授权与管理员管理。",
        "title": "RABC-Go Admin API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/v1/admin/config": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "新增自定义配置项",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "创建配置",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.ConfigCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除自定义配置项，内置配置不允许删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/configs": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "按分组聚合返回全部系统配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "获取系统配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.GetConfigsResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "按分组一次性提交多个配置项的值，整批校验通过才落库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "批量更新配置",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.BatchUpdateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rabc-go_api_apiv1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rabc-go_api_apiv1.BatchUpdateConfigRequest": {
            "type": "object",
            "required": [
                "list"
            ],
            "properties": {
                "list": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/rabc-go_api_apiv1.ConfigUpdateItem"
                    }
                }
            }
        },
        "rabc-go_api_apiv1.ConfigCreateRequest": {
            "type": "object",
            "required": [
                "configGroup",
                "configKey",
                "title",
                "valueType"
            ],
            "properties": {
                "configGroup": {
                    "type": "string",
                    "maxLength": 64,
                    "example": "站点设置"
                },
                "configKey": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "site.name"
                },
                "configValue": {
                    "type": "string",
                    "example": "Nunu Admin"
                },
                "isPublic": {
                    "type": "boolean",
                    "example": false
                },
                "remark": {
                    "type": "string",
                    "maxLength": 255,
                    "example": ""
                },
                "title": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "站点名称"
                },
                "valueType": {
                    "type": "string",
                    "enum": [
                        "string",
                        "int",
                        "bool",
                        "json"
                    ],
                    "example": "string"
                },
                "weight": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "rabc-go_api_apiv1.ConfigGroupItem": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rabc-go_api_apiv1.SysConfigItem"
                    }
                }
            }
        },
        "rabc-go_api_apiv1.ConfigUpdateItem": {
            "type": "object",
            "required": [
                "configKey"
            ],
            "properties": {
                "configKey": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "site.name"
                },
                "configValue": {
                    "type": "string",
                    "example": "Nunu Admin"
                }
            }
        },
...
        "rabc-go_api_apiv1.GetConfigsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/rabc-go_api_apiv1.GetConfigsResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "rabc-go_api_apiv1.GetConfigsResponseData": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rabc-go_api_apiv1.ConfigGroupItem"
                    }
                }
            }
        },
...
        "rabc-go_api_apiv1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rabc-go_api_apiv1.SysConfigItem": {
            "type": "object",
            "properties": {
                "configGroup": {
                    "type": "string"
                },
                "configKey": {
                    "type": "string"
                },
                "configValue": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isPublic": {
                    "type": "boolean"
                },
                "isSystem": {
                    "type": "boolean"
                },
                "remark": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "valueType": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}